semana2/VIERNES/requirements.txt
autopep8==1.7.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
toml==0.10.2
